Enables Hydra on cats build  (#2848)
* Compile Scala with Triplequote Hydra

* Refactoring to workaround initialization issue of package object with inheritance

* added hydra to .gitignore

* Test added hydra to test impact on travis

* fix

* fix

* format

* update hydra

* remove redudant plugin file

* added metrics server config/reformat

* fix metrics conf

* added finite duration instance package / updated hydra

* remove finiteDuration package
