+++ .gitignore
@@ -1,3 +1,4 @@
+.hydra
 project/boot
 target
 .ensime
