+++ hydra-metrics-service.conf
@@ -1,0 +1,17 @@
+# mandatory: include the built-in settings tree
+include "application.conf"
+
+triplequote.dashboard.client {
+  # Optional user ID override
+  # metricsUserId = ${user.name}
+
+  # Optional host ID (e.g. hostname)
+  # metricsHostId = ""
+
+  # Server address to push the metrics data
+  serverUrl = "https://dashboard.triplequote.com/metrics"
+
+  # Optional HTTP basic authentication
+  clientUsername = ${?HYDRA_METRICS_USERNAME}
+  clientPassword = ${?HYDRA_METRICS_PASSWORD}
+}
