+++ kernel/src/main/scala/cats/kernel/instances/FiniteDurationInstances.scala
