+++ kernel/src/main/scala/cats/kernel/instances/UUIDInstances.scala
@@ -1,0 +1,11 @@
+package cats.kernel
+package instances
+
+import java.util.UUID
+
+trait UUIDInstances {
+  implicit val catsKernelStdOrderForUUID: Order[UUID] with Hash[UUID] = new Order[UUID] with Hash[UUID] {
+    def compare(x: UUID, y: UUID): Int = x.compareTo(y)
+    def hash(x: UUID): Int = x.hashCode()
+  }
+}
