+++ kernel/src/main/scala/cats/kernel/instances/boolean/package.scala
@@ -2,25 +2,3 @@ package cats.kernel
 package instances
 
 package object boolean extends BooleanInstances
-
-trait BooleanInstances {
-  implicit val catsKernelStdOrderForBoolean: Order[Boolean] with Hash[Boolean] =
-    new BooleanOrder
-}
-
-class BooleanOrder extends Order[Boolean] with Hash[Boolean] {
-
-  def hash(x: Boolean): Int = x.hashCode()
-  def compare(x: Boolean, y: Boolean): Int =
-    if (x == y) 0 else if (x) 1 else -1
-
-  override def eqv(x: Boolean, y: Boolean): Boolean = x == y
-  override def neqv(x: Boolean, y: Boolean): Boolean = x != y
-  override def gt(x: Boolean, y: Boolean): Boolean = x && !y
-  override def lt(x: Boolean, y: Boolean): Boolean = !x && y
-  override def gteqv(x: Boolean, y: Boolean): Boolean = x == y || x
-  override def lteqv(x: Boolean, y: Boolean): Boolean = x == y || y
-
-  override def min(x: Boolean, y: Boolean): Boolean = x && y
-  override def max(x: Boolean, y: Boolean): Boolean = x || y
-}
